scripts/requirements.txt
boto3>=1.12.14
pyaml>=19.12.0
markdown>=3.2.1
beautifulsoup4>=4.8.2
